small update > new field and info in game.
Removed a console.log.
Added a iteration # label so we know how long it's been going.
Added a label for the slider to tell what it's for and added a small label next to that first label to tell how fast we play turns according to the slider's value.
Added a field where we can set a turn on which to stop the turns from playing. Setting the slider above zero will play until we reach that iteration #.